+++ yandex/data_source_yandex_mdb_mongodb_cluster.go
@@ -197,6 +197,13 @@ func dataSourceYandexMDBMongodbCluster() *schema.Resource {
 										Type:     schema.TypeString,
 										Computed: true,
 									},
+									"roles": {
+										Type:     schema.TypeList,
+										Optional: true,
+										Elem: &schema.Schema{
+											Type: schema.TypeString,
+										},
+									},
 								},
 							},
 						},